Consolidated all cpp programs into a dedicated "cpp_DSA_problem" folder
