+++ cpp_DSA_problems/CountWays.c++
